requirements.txt
python-telegram-bot
openai-whisper
python-dotenv
